implement a templete of routing
